requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
PyYAML==6.0
joblib==1.3.2
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.2.2
gradio==3.41.0
requests==2.31.0